+++ common/src/main/java/net/spell_engine/internals/SpellHelper.java
@@ -587,7 +587,7 @@ public class SpellHelper {
                         }
                         if (attacker instanceof ServerPlayerEntity serverPlayer) {
                             // Map to resolved MeleeAttack structures
-                            var meleeAttacks = Melee.createMeleeAttacks(serverPlayer, attacks, spellId);
+                            var meleeAttacks = Melee.createMeleeAttacks(serverPlayer, attacks, spellEntry);
                             // Send AttackAvailable packet to client
                             var packet = new Packets.AttackAvailable(spellId, meleeAttacks);
                             ServerPlayNetworking.send(serverPlayer, packet);